+++ sources/common/utils.go
@@ -17,10 +17,8 @@ package common
 import (
 	"fmt"
 	"sort"
-	"sync"
 
 	"github.com/GoogleCloudPlatform/spanner-migration-tool/internal"
-	"github.com/GoogleCloudPlatform/spanner-migration-tool/logger"
 	"github.com/GoogleCloudPlatform/spanner-migration-tool/schema"
 	"github.com/GoogleCloudPlatform/spanner-migration-tool/spanner/ddl"
 )
@@ -30,12 +28,6 @@ type UtilsOrderInterface interface {
 	initIndexOrder(conv *internal.Conv)
 }
 type UtilsOrderImpl struct{}
-
-type RunParallelTasksInterface[I any, O any] interface {
-	RunParallelTasks(input []I, numWorkers int, f func(i I, mutex *sync.Mutex) TaskResult[O], fastExit bool) ([]TaskResult[O], error)
-}
-
-type RunParallelTasksImpl[I any, O any] struct{}
 
 // ToNotNull returns true if a column is not nullable and false if it is.
 func ToNotNull(conv *internal.Conv, isNullable string) bool {
@@ -157,68 +149,6 @@ func PrepareValues[T interface{}](conv *internal.Conv, tableId string, colNameId
 		newValues = append(newValues, mapColIdToVal[id])
 	}
 	return newValues, nil
-}
-
-type TaskResult[O any] struct {
-	Result O
-	Err    error
-}
-
-// Run multiple tasks in parallel. The tasks are expected to be thread safe
-// input: 		List of inputs
-// numWorkers: 	Size of worker pool
-// f:			Function to execute the task
-// fastExit: 	If an error is encountered, gracefully exit all running or queued tasks
-// Returns an array of TaskResults and last error
-func (rpt *RunParallelTasksImpl[I, O]) RunParallelTasks(input []I, numWorkers int, f func(i I, mutex *sync.Mutex) TaskResult[O],
-	fastExit bool) ([]TaskResult[O], error) {
-	inputChannel := make(chan I, len(input))
-	outputChannel := make(chan TaskResult[O], len(input))
-
-	wg := &sync.WaitGroup{}
-	defer func() {
-		for range inputChannel {
-			logger.Log.Debug(fmt.Sprint("clearing out pending tasks"))
-		}
-		wg.Wait()
-	}()
-
-	mutex := &sync.Mutex{}
-	logger.Log.Debug(fmt.Sprint("Number of configured workers are ", numWorkers))
-	for w := 0; w < numWorkers; w++ {
-		go processAsync(f, inputChannel, outputChannel, mutex, wg)
-	}
-
-	for _, in := range input {
-		inputChannel <- in
-	}
-	close(inputChannel)
-
-	out := []TaskResult[O]{}
-	for i := 0; i < len(input); i++ {
-		res := <-outputChannel
-		if fastExit && res.Err != nil {
-			logger.Log.Debug(fmt.Sprint("stopping worker pool due to encountered error", res.Err))
-			for range inputChannel {
-				logger.Log.Debug(fmt.Sprint("ignoring task to fast exit"))
-			}
-			wg.Wait()
-			return out, res.Err
-		}
-		out = append(out, res)
-	}
-	logger.Log.Debug(fmt.Sprintf("completed processing of %d tasks", len(out)))
-	return out, nil
-}
-
-func processAsync[I any, O any](f func(i I, mutex *sync.Mutex) TaskResult[O], in chan I,
-	out chan TaskResult[O], mutex *sync.Mutex, wg *sync.WaitGroup) {
-	wg.Add(1)
-	for i := range in {
-		logger.Log.Debug(fmt.Sprint("processing task for input", i))
-		out <- f(i, mutex)
-	}
-	wg.Done()
 }
 
 func ToPGDialectType(standardType ddl.Type, isPk bool) (ddl.Type, []internal.SchemaIssue) {
