+++ sources/common/utils_test.go
@@ -25,6 +25,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	"go.uber.org/zap"
 
+	"github.com/GoogleCloudPlatform/spanner-migration-tool/common/task"
 	"github.com/GoogleCloudPlatform/spanner-migration-tool/internal"
 	"github.com/GoogleCloudPlatform/spanner-migration-tool/logger"
 	"github.com/GoogleCloudPlatform/spanner-migration-tool/schema"
@@ -127,14 +128,14 @@ func TestGetColsAndSchemas(t *testing.T) {
 func TestWorkerPool(t *testing.T) {
 	input := []int{1, 2, 3, 4, 5, 6, 7, 8, 9, 10}
 
-	f := func(i int, mutex *sync.Mutex) TaskResult[int] {
+	f := func(i int, mutex *sync.Mutex) task.TaskResult[int] {
 		sleepTime := time.Duration(rand.Intn(1000 * 1000))
 		time.Sleep(sleepTime)
-		res := TaskResult[int]{Result: i, Err: nil}
+		res := task.TaskResult[int]{Result: i, Err: nil}
 		return res
 	}
 
-	r := RunParallelTasksImpl[int, int]{}
+	r := task.RunParallelTasksImpl[int, int]{}
 	out, _ := r.RunParallelTasks(input, 5, f, false)
 	assert.Equal(t, len(input), len(out), fmt.Sprintln("jobs not processed"))
 }
